Delete the handwritten Message type

+++ google-cloud-pubsub/src/test/java/com/google/cloud/pubsub/PublisherImplTest.java
@@ -407,7 +407,10 @@ public class PublisherImplTest {
     }
     try {
       builder.setBundlingSettings(
-          Publisher.DEFAULT_BUNDLING_SETTINGS.toBuilder().setRequestByteThreshold(null).build());
+          Publisher.DEFAULT_BUNDLING_SETTINGS
+              .toBuilder()
+              .setRequestByteThreshold((Long) null)
+              .build());
       fail("Should have thrown an NullPointerException");
     } catch (NullPointerException expected) {
       // Expected
@@ -451,7 +454,10 @@ public class PublisherImplTest {
         Publisher.DEFAULT_BUNDLING_SETTINGS.toBuilder().setElementCountThreshold(1).build());
     try {
       builder.setBundlingSettings(
-          Publisher.DEFAULT_BUNDLING_SETTINGS.toBuilder().setElementCountThreshold(null).build());
+          Publisher.DEFAULT_BUNDLING_SETTINGS
+              .toBuilder()
+              .setElementCountThreshold((Long) null)
+              .build());
       fail("Should have thrown an NullPointerException");
     } catch (NullPointerException expected) {
       // Expected
